# Conflicts:
#	src/main/java/com/example/bookangel/beans/vo/PaymentVO.java

+++ src/main/java/com/example/bookangel/beans/vo/PaymentVO.java
@@ -6,14 +6,13 @@ import org.springframework.stereotype.Component;
 @Component
 @Data
 public class PaymentVO {
-
-    private int paymentNum;
-    private Long memberNum;
-    private int couponNum;
+    private long paymentNum;
+    private long memberNum;
+    private long couponNum;
     private String impUid;
     private String merchantUid;
-    private int paymentType;
-    private int subMonth;
+    private long paymentType;
+    private long subMonth;
     private String subDate;
     private String expireDate;
 }
